+++ zdsimScanner55/Core/Form2.Core.AxisMarks.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Collections.Specialized;
+using System.Windows.Forms;
+
+namespace zdsimScanner
+{
+    public partial class Form2 : Form
+    {
+        // сюда переносим поля/методы по теме файла
+
+        private void __partial_test()
+        {
+            this.Text = "OK";
+        }
+
+
+    }
+}
