+++ zdsimScanner55/Core/Form2.Core.Persistence.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Collections.Specialized;
+using System.Windows.Forms;
+
+namespace zdsimScanner
+{
+    public partial class Form2 : Form
+    {
+        // сюда переносим поля/методы по теме файла
+    }
+}
