+++ zdsimScanner55/zdsimScanner.csproj
@@ -119,6 +119,21 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Core\Form2.Core.AxisMarks.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Core\Form2.Core.Persistence.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Core\Form2.Core.ZdSim.LocoResetActions.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Core\Form2.Core.ZdSim.RowBindings.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Core\Form2.Core.SettingsNormalize.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="Core\LocoMemoryHelpers.cs" />
     <Compile Include="Core\ProcessMemory.cs" />
     <Compile Include="Form1.cs">
@@ -129,11 +144,26 @@
       <DependentUpon>Form1.cs</DependentUpon>
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Compile>
+    <Compile Include="Core\Form2.Core.ZdSim.SoundBindings.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="Form2.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Form2.Designer.cs">
       <DependentUpon>Form2.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Events\Form2.Events.AxisMarks.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Events\Form2.Events.Common.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Events\Form2.Events.GlobalReset.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Events\Form2.Events.ZdSim.cs">
+      <SubType>Form</SubType>
     </Compile>
     <Compile Include="Form3.cs">
       <SubType>Form</SubType>
